Lazy load rpc-request to avoid unnecessary dependencies when using a different implementation.
Previously, if you set a custom RPC request handler, the es6-promise and superagent dependencies were still required and an exception would be thrown if they are missing during runtime. This change lazy-loads the default RPC request handler so errors won't be thrown about missing dependencies if the handler implementation is changed.