Error handling in case of timeout
If the error code is ETIMEDOUT will response be undefined. When response.text is called in buildCustomError will the server crash and output:

error: response.text,
TypeError: Cannot read property 'text' of undefined

By checking if response is undefined is it possible to give proper error handling and not crash the server.